+++ Waypoint/Services/SearchCompleterService.swift
@@ -1,0 +1,31 @@
+import MapKit
+import Observation
+
+@Observable
+final class SearchCompleterService: NSObject, MKLocalSearchCompleterDelegate {
+    private(set) var suggestions: [MKLocalSearchCompletion] = []
+
+    private let completer = MKLocalSearchCompleter()
+
+    override init() {
+        super.init()
+        completer.delegate = self
+        completer.resultTypes = [.address, .pointOfInterest]
+    }
+
+    func updateQuery(_ fragment: String) {
+        completer.queryFragment = fragment
+    }
+
+    func updateRegion(_ region: MKCoordinateRegion) {
+        completer.region = region
+    }
+
+    func completerDidUpdateResults(_ completer: MKLocalSearchCompleter) {
+        suggestions = completer.results
+    }
+
+    func completer(_ completer: MKLocalSearchCompleter, didFailWithError error: Error) {
+        suggestions = []
+    }
+}